+++ src/commands/expressions/update/add.ts
@@ -1,7 +1,7 @@
 import type { NativeNumber, NativeSet } from "../../../types.js";
 import type { AttributeNames } from "../../attributes/names.js";
 import type { AttributeValues } from "../../attributes/values.js";
-import type { AttributeOperand } from "../operands/name.js";
+import type { PathOperand } from "../operands/name.js";
 import type { ValueOperand } from "../operands/value.js";
 import type { IUpdateAction, UpdateAction } from "./action.js";
 import type { UpdateExpressionClauses } from "./clauses.js";
@@ -10,11 +10,11 @@ type NumberOrSet = NativeNumber | NativeSet;
 
 // Note: the first operand *must* be an attribute name, and the second operand *must* be a value.
 export class AddAction implements IUpdateAction {
-  private readonly path: AttributeOperand;
+  private readonly path: PathOperand;
   private readonly value: ValueOperand<NumberOrSet>;
 
   private constructor(params: {
-    path: AttributeOperand;
+    path: PathOperand;
     value: ValueOperand<NumberOrSet>;
   }) {
     const { path, value } = params;
@@ -35,7 +35,7 @@ export class AddAction implements IUpdateAction {
   }
 
   static from(params: {
-    path: AttributeOperand;
+    path: PathOperand;
     value: ValueOperand<NumberOrSet>;
   }): AddAction {
     return new AddAction(params);
@@ -60,7 +60,7 @@ export class AddAction implements IUpdateAction {
  */
 
 export function add(
-  path: AttributeOperand,
+  path: PathOperand,
   value: ValueOperand<NumberOrSet>,
 ): UpdateAction {
   return AddAction.from({ path, value });