+++ src/commands/expressions/update/delete.ts
@@ -1,17 +1,17 @@
 import type { NativeSet } from "../../../types.js";
 import type { AttributeNames } from "../../attributes/names.js";
 import type { AttributeValues } from "../../attributes/values.js";
-import type { AttributeOperand } from "../operands/name.js";
+import type { PathOperand } from "../operands/name.js";
 import type { ValueOperand } from "../operands/value.js";
 import type { IUpdateAction, UpdateAction } from "./action.js";
 import type { UpdateExpressionClauses } from "./clauses.js";
 
 export class DeleteAction implements IUpdateAction {
-  private readonly path: AttributeOperand;
+  private readonly path: PathOperand;
   private readonly value: ValueOperand<NativeSet>;
 
   private constructor(params: {
-    path: AttributeOperand;
+    path: PathOperand;
     value: ValueOperand<NativeSet>;
   }) {
     const { path, value } = params;
@@ -34,7 +34,7 @@ export class DeleteAction implements IUpdateAction {
   }
 
   static from(params: {
-    path: AttributeOperand;
+    path: PathOperand;
     value: ValueOperand<NativeSet>;
   }): DeleteAction {
     return new DeleteAction(params);
@@ -56,7 +56,7 @@ export class DeleteAction implements IUpdateAction {
  */
 
 export function deleteFrom(
-  path: AttributeOperand,
+  path: PathOperand,
   value: ValueOperand<NativeSet>,
 ): UpdateAction {
   return DeleteAction.from({ path, value });