+++ src/commands/expressions/update/remove.ts
@@ -1,12 +1,12 @@
 import type { AttributeNames } from "../../attributes/names.js";
-import type { AttributeOperand } from "../operands/name.js";
+import type { PathOperand } from "../operands/name.js";
 import type { IUpdateAction, UpdateAction } from "./action.js";
 import type { UpdateExpressionClauses } from "./clauses.js";
 
 export class RemoveAction implements IUpdateAction {
-  private readonly path: AttributeOperand;
+  private readonly path: PathOperand;
 
-  private constructor(path: AttributeOperand) {
+  private constructor(path: PathOperand) {
     this.path = path;
   }
 
@@ -19,7 +19,7 @@ export class RemoveAction implements IUpdateAction {
     return this.path.substitute({ names });
   }
 
-  static from(path: AttributeOperand): RemoveAction {
+  static from(path: PathOperand): RemoveAction {
     return new RemoveAction(path);
   }
 }
@@ -33,6 +33,6 @@ export class RemoveAction implements IUpdateAction {
  *
  * @see https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/Expressions.UpdateExpressions.html#Expressions.UpdateExpressions.DELETE
  */
-export function remove(path: AttributeOperand): UpdateAction {
+export function remove(path: PathOperand): UpdateAction {
   return RemoveAction.from(path);
 }