+++ test/unit/commands/expressions/condition.spec.ts
@@ -3,10 +3,10 @@ import { expect } from "@infra-blocks/test";
 import { AttributeNames } from "../../../../src/commands/attributes/names.js";
 import { AttributeValues } from "../../../../src/commands/attributes/values.js";
 import {
-  AttributeOperand,
-  attribute,
   Condition,
   not,
+  PathOperand,
+  path,
   size,
   ValueOperand,
   value,
@@ -33,12 +33,12 @@ describe("commands.expressions.condition", () => {
   }
 
   // TODO: test manually how to reference `.` paths (i.e is it #path or #path.#subpath)
-  describe(AttributeOperand.name, () => {
+  describe(PathOperand.name, () => {
     describe("beginsWith", () => {
       it("should work with attribute rhs", () => {
         const lhs = "test.attribute.lhs";
         const rhs = "test.attribute.rhs";
-        const expression = where(attribute(lhs)).beginsWith(attribute(rhs));
+        const expression = where(path(lhs)).beginsWith(path(rhs));
         const { match, names } = conditionMatch({
           condition: expression,
           matcher: /begins_with\((#\S+),\s*(#\S+)\)/,
@@ -52,7 +52,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "test.attribute.lhs";
         const rhs = "I am a value";
         const { match, names, values } = conditionMatch({
-          condition: where(attribute(lhs)).beginsWith(value(rhs)),
+          condition: where(path(lhs)).beginsWith(value(rhs)),
           matcher: /begins_with\((#\S+),\s*(:\S+)\)/,
         });
         const lhsSubstitution = match[1];
@@ -67,10 +67,7 @@ describe("commands.expressions.condition", () => {
         const lower = "test.attribute.lower";
         const upper = "test.attribute.upper";
         const { match, names } = conditionMatch({
-          condition: where(attribute(lhs)).between(
-            attribute(lower),
-            attribute(upper),
-          ),
+          condition: where(path(lhs)).between(path(lower), path(upper)),
           matcher: /(#\S+)\s+BETWEEN\s+(#\S+)\s+AND\s+(#\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -85,7 +82,7 @@ describe("commands.expressions.condition", () => {
         const lower = "I am a value";
         const upper = "I am also a value";
         const { match, names, values } = conditionMatch({
-          condition: where(attribute(lhs)).between(value(lower), value(upper)),
+          condition: where(path(lhs)).between(value(lower), value(upper)),
           matcher: /(#\S+)\s+BETWEEN\s+(:\S+)\s+AND\s+(:\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -101,7 +98,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "test.attribute.lhs";
         const rhs = "test.attribute.rhs";
         const { match, names } = conditionMatch({
-          condition: where(attribute(lhs)).contains(attribute(rhs)),
+          condition: where(path(lhs)).contains(path(rhs)),
           matcher: /contains\((#\S+),\s*(#\S+)\)/,
         });
         const lhsSubstitution = match[1];
@@ -113,7 +110,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "test.attribute.lhs";
         const rhs = "I am a value";
         const { match, names, values } = conditionMatch({
-          condition: where(attribute(lhs)).contains(value(rhs)),
+          condition: where(path(lhs)).contains(value(rhs)),
           matcher: /contains\((#\S+),\s*(:\S+)\)/,
         });
         const lhsSubstitution = match[1];
@@ -127,7 +124,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "test.attribute.lhs";
         const rhs = "test.attribute.rhs";
         const { match, names } = conditionMatch({
-          condition: where(attribute(lhs)).equals(attribute(rhs)),
+          condition: where(path(lhs)).equals(path(rhs)),
           matcher: /(#\S+)\s*=\s*(#\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -139,7 +136,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "test.attribute.lhs";
         const rhs = "I am a value";
         const { match, names, values } = conditionMatch({
-          condition: where(attribute(lhs)).equals(value(rhs)),
+          condition: where(path(lhs)).equals(value(rhs)),
           matcher: /(#\S+)\s*=\s*(:\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -151,7 +148,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "test.attribute.lhs";
         const rhs = "I contain things";
         const { match, names, values } = conditionMatch({
-          condition: where(attribute(lhs)).equals(size(value(rhs))),
+          condition: where(path(lhs)).equals(size(value(rhs))),
           matcher: /(#\S+)\s*=\s*size\((:\S+)\)/,
         });
         const lhsSubstitution = match[1];
@@ -164,7 +161,7 @@ describe("commands.expressions.condition", () => {
       it("should work", () => {
         const name = "test.attribute";
         const { match, names } = conditionMatch({
-          condition: where(attribute(name)).exists(),
+          condition: where(path(name)).exists(),
           matcher: /attribute_exists\((#.+)\)/,
         });
         const substitution = match[1];
@@ -176,7 +173,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "test.attribute.lhs";
         const rhs = "test.attribute.rhs";
         const { match, names } = conditionMatch({
-          condition: where(attribute(lhs)).gt(attribute(rhs)),
+          condition: where(path(lhs)).gt(path(rhs)),
           matcher: /(#\S+)\s*>\s*(#\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -188,7 +185,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "test.attribute.lhs";
         const rhs = "I am a value";
         const { match, names, values } = conditionMatch({
-          condition: where(attribute(lhs)).gt(value(rhs)),
+          condition: where(path(lhs)).gt(value(rhs)),
           matcher: /(#\S+)\s*>\s*(:\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -202,7 +199,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "test.attribute.lhs";
         const rhs = "test.attribute.rhs";
         const { match, names } = conditionMatch({
-          condition: where(attribute(lhs)).gte(attribute(rhs)),
+          condition: where(path(lhs)).gte(path(rhs)),
           matcher: /(#\S+)\s*>=\s*(#\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -214,7 +211,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "test.attribute.lhs";
         const rhs = "I am a value";
         const { match, names, values } = conditionMatch({
-          condition: where(attribute(lhs)).gte(value(rhs)),
+          condition: where(path(lhs)).gte(value(rhs)),
           matcher: /(#\S+)\s*>=\s*(:\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -226,15 +223,15 @@ describe("commands.expressions.condition", () => {
     describe("in", () => {
       it("should throw when no operands are provided", () => {
         expect(() => {
-          where(attribute("test.attribute.lhs")).in();
+          where(path("test.attribute.lhs")).in();
         }).to.throw();
       });
       it("should throw when more than 100 operands are provided", () => {
         const operands = Array.from({ length: 101 }, (_, i) =>
-          attribute(`test.attribute.${i}`),
+          path(`test.attribute.${i}`),
         );
         expect(() => {
-          where(attribute("test.attribute.lhs")).in(...operands);
+          where(path("test.attribute.lhs")).in(...operands);
         }).to.throw();
       });
       it("should work with attribute operands", () => {
@@ -245,9 +242,7 @@ describe("commands.expressions.condition", () => {
           "test.attribute.operand3",
         ];
         const { match, names } = conditionMatch({
-          condition: where(attribute(lhs)).in(
-            ...operands.map((op) => attribute(op)),
-          ),
+          condition: where(path(lhs)).in(...operands.map((op) => path(op))),
           matcher: /(#\S+)\s+IN\s+\((#\S+),(#\S+),(#\S+)\)/,
         });
         const lhsSubstitution = match[1];
@@ -267,9 +262,7 @@ describe("commands.expressions.condition", () => {
           "I am yet another",
         ];
         const { match, names, values } = conditionMatch({
-          condition: where(attribute(lhs)).in(
-            ...operands.map((op) => value(op)),
-          ),
+          condition: where(path(lhs)).in(...operands.map((op) => value(op))),
           matcher: /(#\S+)\s+IN\s+\((:\S+),(:\S+),(:\S+)\)/,
         });
         const lhsSubstitution = match[1];
@@ -287,7 +280,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "test.attribute.lhs";
         const rhs = "test.attribute.rhs";
         const { match, names } = conditionMatch({
-          condition: where(attribute(lhs)).lt(attribute(rhs)),
+          condition: where(path(lhs)).lt(path(rhs)),
           matcher: /(#\S+)\s*<\s*(#\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -299,7 +292,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "test.attribute.lhs";
         const rhs = "I am a value";
         const { match, names, values } = conditionMatch({
-          condition: where(attribute(lhs)).lt(value(rhs)),
+          condition: where(path(lhs)).lt(value(rhs)),
           matcher: /(#\S+)\s*<\s*(:\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -313,7 +306,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "test.attribute.lhs";
         const rhs = "test.attribute.rhs";
         const { match, names } = conditionMatch({
-          condition: where(attribute(lhs)).lte(attribute(rhs)),
+          condition: where(path(lhs)).lte(path(rhs)),
           matcher: /(#\S+)\s*<=\s*(#\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -325,7 +318,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "test.attribute.lhs";
         const rhs = "I am a value";
         const { match, names, values } = conditionMatch({
-          condition: where(attribute(lhs)).lte(value(rhs)),
+          condition: where(path(lhs)).lte(value(rhs)),
           matcher: /(#\S+)\s*<=\s*(:\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -339,7 +332,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "test.attribute.lhs";
         const rhs = "test.attribute.rhs";
         const { match, names } = conditionMatch({
-          condition: where(attribute(lhs)).ne(attribute(rhs)),
+          condition: where(path(lhs)).ne(path(rhs)),
           matcher: /(#\S+)\s*<>\s*(#\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -351,7 +344,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "test.attribute.lhs";
         const rhs = "I am a value";
         const { match, names, values } = conditionMatch({
-          condition: where(attribute(lhs)).ne(value(rhs)),
+          condition: where(path(lhs)).ne(value(rhs)),
           matcher: /(#\S+)\s*<>\s*(:\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -362,13 +355,13 @@ describe("commands.expressions.condition", () => {
     });
     describe("notExists", () => {
       it("should work", () => {
-        const path = "test.attribute";
+        const attribute = "test.attribute";
         const { match, names } = conditionMatch({
-          condition: where(attribute(path)).notExists(),
+          condition: where(path(attribute)).notExists(),
           matcher: /attribute_not_exists\((#.+)\)/,
         });
         const substitution = match[1];
-        expect(substitution).to.equal(names.substitute(path));
+        expect(substitution).to.equal(names.substitute(attribute));
       });
     });
     describe("type", () => {
@@ -376,7 +369,7 @@ describe("commands.expressions.condition", () => {
         const name = "test.attribute";
         const type = "S";
         const { match, names, values } = conditionMatch({
-          condition: where(attribute(name)).isType(value(type)),
+          condition: where(path(name)).isType(value(type)),
           matcher: /attribute_type\((#.+),\s*(:.+)\)/,
         });
         const substitution = match[1];
@@ -392,7 +385,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "I am a value";
         const rhs = "test.attribute.rhs";
         const { match, names, values } = conditionMatch({
-          condition: where(value(lhs)).beginsWith(attribute(rhs)),
+          condition: where(value(lhs)).beginsWith(path(rhs)),
           matcher: /begins_with\((:\S+),\s*(#\S+)\)/,
         });
         const lhsSubstitution = match[1];
@@ -419,10 +412,7 @@ describe("commands.expressions.condition", () => {
         const lower = "test.attribute.lower";
         const upper = "test.attribute.upper";
         const { match, names, values } = conditionMatch({
-          condition: where(value(lhs)).between(
-            attribute(lower),
-            attribute(upper),
-          ),
+          condition: where(value(lhs)).between(path(lower), path(upper)),
           matcher: /(:\S+)\s+BETWEEN\s+(#\S+)\s+AND\s+(#\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -453,7 +443,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "I am a value";
         const rhs = "test.attribute.rhs";
         const { match, names, values } = conditionMatch({
-          condition: where(value(lhs)).contains(attribute(rhs)),
+          condition: where(value(lhs)).contains(path(rhs)),
           matcher: /contains\((:\S+),\s*(#\S+)\)/,
         });
         const lhsSubstitution = match[1];
@@ -479,7 +469,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "I am a value";
         const rhs = "test.attribute.rhs";
         const { match, names, values } = conditionMatch({
-          condition: where(value(lhs)).eq(attribute(rhs)),
+          condition: where(value(lhs)).eq(path(rhs)),
           matcher: /(:\S+)\s*=\s*(#\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -517,7 +507,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "I am a value";
         const rhs = "test.attribute.rhs";
         const { match, names, values } = conditionMatch({
-          condition: where(value(lhs)).gt(attribute(rhs)),
+          condition: where(value(lhs)).gt(path(rhs)),
           matcher: /(:\S+)\s*>\s*(#\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -543,7 +533,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "I am a value";
         const rhs = "test.attribute.rhs";
         const { match, names, values } = conditionMatch({
-          condition: where(value(lhs)).gte(attribute(rhs)),
+          condition: where(value(lhs)).gte(path(rhs)),
           matcher: /(:\S+)\s*>=\s*(#\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -586,9 +576,7 @@ describe("commands.expressions.condition", () => {
           "test.attribute.operand3",
         ];
         const { match, names, values } = conditionMatch({
-          condition: where(value(lhs)).in(
-            ...operands.map((op) => attribute(op)),
-          ),
+          condition: where(value(lhs)).in(...operands.map((op) => path(op))),
           matcher: /(:\S+)\s+IN\s+\((#\S+),(#\S+),(#\S+)\)/,
         });
         const lhsSubstitution = match[1];
@@ -626,7 +614,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "I am a value";
         const rhs = "test.attribute.rhs";
         const { match, names, values } = conditionMatch({
-          condition: where(value(lhs)).lt(attribute(rhs)),
+          condition: where(value(lhs)).lt(path(rhs)),
           matcher: /(:\S+)\s*<\s*(#\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -652,7 +640,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "I am a value";
         const rhs = "test.attribute.rhs";
         const { match, names, values } = conditionMatch({
-          condition: where(value(lhs)).lte(attribute(rhs)),
+          condition: where(value(lhs)).lte(path(rhs)),
           matcher: /(:\S+)\s*<=\s*(#\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -678,7 +666,7 @@ describe("commands.expressions.condition", () => {
         const lhs = "I am a value";
         const rhs = "test.attribute.rhs";
         const { match, names, values } = conditionMatch({
-          condition: where(value(lhs)).ne(attribute(rhs)),
+          condition: where(value(lhs)).ne(path(rhs)),
           matcher: /(:\S+)\s*<>\s*(#\S+)/,
         });
         const lhsSubstitution = match[1];
@@ -705,8 +693,8 @@ describe("commands.expressions.condition", () => {
       it("should properly combine two expressions", () => {
         const left = "attr.left";
         const right = "attr.right";
-        const leftCondition = where(attribute(left)).exists();
-        const rightCondition = where(attribute(right)).notExists();
+        const leftCondition = where(path(left)).exists();
+        const rightCondition = where(path(right)).notExists();
         const { match, names } = conditionMatch({
           condition: leftCondition.and(rightCondition),
           matcher:
@@ -722,8 +710,8 @@ describe("commands.expressions.condition", () => {
       it("should properly combine two expressions", () => {
         const left = "attr.left";
         const right = "attr.right";
-        const leftCondition = where(attribute(left)).exists();
-        const rightCondition = where(attribute(right)).notExists();
+        const leftCondition = where(path(left)).exists();
+        const rightCondition = where(path(right)).notExists();
         const { match, names } = conditionMatch({
           condition: leftCondition.or(rightCondition),
           matcher:
@@ -740,7 +728,7 @@ describe("commands.expressions.condition", () => {
     it("should properly negate the expression", () => {
       const name = "test.attribute";
       const { match, names } = conditionMatch({
-        condition: not(where(attribute(name)).exists()),
+        condition: not(where(path(name)).exists()),
         matcher: /NOT \(attribute_exists\((#.+)\)\)/,
       });
       const substitution = match[1];