+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.26.0] - 2025-07-28
+
+### Changed
+
+- Renamed `AttributeOperand` for `PathOperand`. It's less ambiguous. The factory
+function has also been renamed from `attribute` to `path`.
+
 ## [0.25.0] - 2025-07-26
 
 ### Changed
@@ -337,6 +344,7 @@ intuitive.
 - Initial release of the package! Move the implementation work in progress from another
 project to here.
 
+[0.26.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.25.0...v0.26.0
 [0.25.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.24.0...v0.25.0
 [0.24.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.23.0...v0.24.0
 [0.23.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.22.0...v0.23.0